Update README.md
I have some problems with this code. It is an arc or similar to a line that has a slope, and I wanted to connect it to two countries inside the globe, but when I do that, the matter gets distorted. As soon as I move the globe or even zoom in or out, the size of the shape decreases and moves, so I want it to be fixed  ?